+++ src/App.tsx
@@ -3,6 +3,18 @@
 import React, { useEffect, useState } from 'react';
 import { Todo } from './types/Todo';
 import { getTodos } from './api/todos';
+import { AppHeader } from './app-header';
+import { Error } from './error';
+import { AppFooter } from './app-footer';
+import { ToDo } from './todo';
+
+enum Errors {
+  'errorLoad' = 'Unable to load todos',
+  'errorEmptyTitle' = 'Title should not be empty',
+  'errorAdd' = 'Unable to add a todo',
+  'errorDelete' = 'Unable to delete a todo',
+  'errorUpdate' = 'Unable to update a todo',
+}
 
 export const App: React.FC = () => {
   const [todos, setTodos] = useState<Todo[] | null>(null);
@@ -22,160 +34,34 @@ export const App: React.FC = () => {
     return true;
   });
 
-  const completedTodos = todos?.filter(todo => !todo.completed).length;
-
   useEffect(() => {
     setError(null);
     setIsLoading(true);
 
     getTodos()
       .then(res => setTodos(res))
-      .catch(() => setError('Unable to load todos'))
+      .catch(() => setError(Errors.errorLoad))
       .finally(() => setIsLoading(false));
   }, []);
-
-  useEffect(() => {
-    if (!error) {
-      return;
-    }
-
-    const timeout = setTimeout(() => setError(null), 3000);
-
-    return () => clearTimeout(timeout);
-  }, [error]);
 
   return (
     <div className="todoapp">
       <h1 className="todoapp__title">todos</h1>
-
+      <AppHeader />
       <div className="todoapp__content">
-        <header className="todoapp__header">
-          <button
-            type="button"
-            className="todoapp__toggle-all active"
-            data-cy="ToggleAllButton"
-            onClick={() => {}}
-          />
-
-          {/* Add a todo on form submit */}
-          <form onSubmit={() => {}}>
-            <input
-              data-cy="NewTodoField"
-              type="text"
-              className="todoapp__new-todo"
-              placeholder="What needs to be done?"
-            />
-          </form>
-        </header>
-
         {!isLoading && (
           <section className="todoapp__main" data-cy="TodoList">
-            {filteredTodos?.map(todo => (
-              <div
-                data-cy="Todo"
-                key={todo.id}
-                className={`todo ${todo.completed ? 'completed' : ''}`}
-              >
-                <label className="todo__status-label">
-                  <input
-                    data-cy="TodoStatus"
-                    type="checkbox"
-                    className="todo__status"
-                    checked={todo.completed}
-                    onChange={() => toggleTodo(todo.id)}
-                  />
-                </label>
-
-                <span data-cy="TodoTitle" className="todo__title">
-                  {todo.title}
-                </span>
-
-                {/* Remove button appears only on hover */}
-                <button
-                  type="button"
-                  className="todo__remove"
-                  data-cy="TodoDelete"
-                  onClick={() => {}}
-                >
-                  ×
-                </button>
-
-                {/* overlay will cover the todo while it is being deleted or updated */}
-                <div data-cy="TodoLoader" className="modal overlay">
-                  <div className="modal-background has-background-white-ter" />
-                  <div className="loader" />
-                </div>
-              </div>
-            ))}
+            {filteredTodos?.map(todo => <ToDo todo={todo} key={todo.id} />)}
           </section>
         )}
 
         {/* Hide the footer if there are no todos */}
         {todos?.length && (
-          <footer
-            className={`todoapp__footer ${!todos?.length ? 'hidden' : ''}`}
-            data-cy="Footer"
-          >
-            <span className="todo-count" data-cy="TodosCounter">
-              {`${completedTodos} items left`}
-            </span>
-
-            {/* Active link should have the 'selected' class */}
-            <nav className="filter" data-cy="Filter">
-              <a
-                href="#/"
-                className={`filter__link ${filter === 'all' ? 'selected' : ''}`}
-                data-cy="FilterLinkAll"
-                onClick={() => setFilter('all')}
-              >
-                All
-              </a>
-
-              <a
-                href="#/active"
-                className={`filter__link ${filter === 'active' ? 'selected' : ''}`}
-                data-cy="FilterLinkActive"
-                onClick={() => setFilter('active')}
-              >
-                Active
-              </a>
-
-              <a
-                href="#/completed"
-                className={`filter__link ${filter === 'completed' ? 'selected' : ''}`}
-                data-cy="FilterLinkCompleted"
-                onClick={() => setFilter('completed')}
-              >
-                Completed
-              </a>
-            </nav>
-
-            {/* this button should be disabled if there are no completed todos */}
-            <button
-              type="button"
-              className="todoapp__clear-completed"
-              data-cy="ClearCompletedButton"
-            >
-              Clear completed
-            </button>
-          </footer>
+          <AppFooter todos={todos} filter={filter} setFilter={setFilter} />
         )}
       </div>
 
-      {/* DON'T use conditional rendering to hide the notification */}
-      {/* Add the 'hidden' class to hide the message smoothly */}
-      <div
-        data-cy="ErrorNotification"
-        className={`notification is-danger is-light has-text-weight-normal ${!error ? 'hidden' : ''}`}
-      >
-        <button
-          data-cy="HideErrorButton"
-          type="button"
-          className="delete"
-          onClick={() => setError(null)}
-        />
-        {error}
-      </div>
+      <Error error={error} setError={setError} />
     </div>
   );
 };
