+++ src/app-header.tsx
@@ -1,0 +1,22 @@
+export const AppHeader = () => {
+  return (
+    <header className="todoapp__header">
+      <button
+        type="button"
+        className="todoapp__toggle-all active"
+        data-cy="ToggleAllButton"
+        onClick={() => {}}
+      />
+
+      {/* Add a todo on form submit */}
+      <form onSubmit={() => {}}>
+        <input
+          data-cy="NewTodoField"
+          type="text"
+          className="todoapp__new-todo"
+          placeholder="What needs to be done?"
+        />
+      </form>
+    </header>
+  );
+};
