+++ src/todo.tsx
@@ -1,0 +1,43 @@
+import { Todo } from './types/Todo';
+
+export const ToDo = ({ todo, key }: { todo: Todo; key: number }) => {
+  return (
+    <div
+      data-cy="Todo"
+      key={key}
+      className={`todo ${todo.completed ? 'completed' : ''}`}
+    >
+      <label className="todo__status-label" htmlFor={`todo-status-${todo.id}`}>
+        <input
+          id={`todo-status-${todo.id}`}
+          data-cy="TodoStatus"
+          type="checkbox"
+          className="todo__status"
+          checked={todo.completed}
+          aria-label="Toggle todo status"
+          onChange={() => {}}
+        />
+      </label>
+
+      <span data-cy="TodoTitle" className="todo__title">
+        {todo.title}
+      </span>
+
+      {/* Remove button appears only on hover */}
+      <button
+        type="button"
+        className="todo__remove"
+        data-cy="TodoDelete"
+        onClick={() => {}}
+      >
+        ×
+      </button>
+
+      {/* overlay will cover the todo while it is being deleted or updated */}
+      <div data-cy="TodoLoader" className="modal overlay">
+        <div className="modal-background has-background-white-ter" />
+        <div className="loader" />
+      </div>
+    </div>
+  );
+};
